Code organization and rewrite
This rewrite moves individual code pieces into modules, leaving
only orchestration logic in the CLI module.

It also spread retries where they are really necessary and manages
the Gemini uploads to remove files after they are done.

Finally, it adds better processing for audio streams that are not
in any of the supported formats.